requirements.txt
youtube_transcript_api
deepmultilingualpunctuation
openai 
FastAPI
uvicorn 
python-dotenv
streamlit
